pyproject.toml
[project]
name = "langbot-app"
version = "1.0.0"
description = "A Local, Privacy-First Chatbot Application powered by LangChain and Ollama (Phi), enabling fast, offline LLM interactions without relying on cloud APIs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-ollama>=0.3.6",
    "langserve>=0.3.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "streamlit>=1.47.1",
    "uvicorn>=0.36.0",
]
